pyproject.toml
[project]
name = "pbapi"
version = "0.1.0"
description = "Plant-Based API"
requires-python = ">=3.12"
dependencies = [
    "alembic",
    "doppler-sdk",
    "appwrite",
    "sqlalchemy",
    "pydantic[email]",
    "python-dotenv",
    "psycopg2>=2.9.11",
    "fastapi>=0.128.0",
]

[project.optional-dependencies]
dev = [
    "black",
    "coverage",
    "pylint",
    "pytest",
    "pytest-cov",
    "testcontainers[postgres]",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "pylint>=4.0.4",
    "uvicorn>=0.40.0",
]
